// front-end/src/api/model/createLocationDtoType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Wardrobe API
 * The Wardrobe App API description
 * OpenAPI spec version: 1.0
 */

export type CreateLocationDtoType = typeof CreateLocationDtoType[keyof typeof CreateLocationDtoType];


export const CreateLocationDtoType = {
  LOCATION: 'LOCATION',
  HOUSE: 'HOUSE',
  ROOM: 'ROOM',
  CABINET: 'CABINET',
  SHELF: 'SHELF',
  BOX: 'BOX',
  CLOSET: 'CLOSET',
  SECTION: 'SECTION',
  DRESSER: 'DRESSER',
  DRAWER: 'DRAWER',
} as const;